Rename Scylla directory and fix vars
