+++ example-scylla/public/images/logo-graphic-white.png
