+++ example-scylla/public/javascripts/main.js
